Dragon Scales, CliLocFunctions, and Global Player
- CliLocFunctions now parses a check for a single item with no greater
amount and its weight being 255. If so, the CliLoc won't show the
weight.
- Added a few global events to keep single items of weight 255 from
being picked up. This is per EA behaviors as they keep an unseen weight
limit a single item can have that a char can lift 255 is their default
TileData setting to stop this from occuring.
- Added in DragonScales. They don't have their base item attached yet,
but should by the morning.