Minor Update
- SORRY! Forgot to add in ID's on shields and forgot to add in the
Gargish Wooden Shield! Done!
- Trimmed monster files some more and moved Fire Gargoyle Names to Names
file.
- Scripted the Scale Shield to use Metal Kite Shield DispID when
equipped and switch back when not equipped. (It has Metal Kite Shield
weight and info in Art.Mul) (Close facsimile in appearance as well).
- Added Dupre's Shield to Artifacts Other. Honestly, all artifacts
should be hand coded to not run functions for checks on
Equip/Unequip/ClientTooltip. This will save a good bit of memory as the
props are static. (Reminder to upload my backwards compat functions,
making the props I did for hard code are a bit harder to remember and
takes a lot of cross referencing. So I did strict functions on my end
(Sorry, it was all to be experimental when I added it)).
- Removed unneeded comment in CliLoc Functions.