StartingItems
- Okay someone changed some things in source and until Xun does his
merge, it would be almost silly for me to fix what they changed. So
right now Default_Male/Female or whatever is not taking race selectors
like it should. Gender selectors seem to be missing as well on the
newbie templates. Can add this in when Xun finishes his merge. Sent him
a fix for default gender stuff. Might not add back in the third param.
Not sure yet. If default male/female work fine... it probably won't be
needed.