Doors update
Note: There will be one more after this, sorry. ML doors are a bit of a
pain and we need to double check the 0,0 offset.
- Added several more door types to the file.
- Offsets are placed on all doors now with proper DispID to change into.
- All that is left is ML doors, Shadow and Crystal, and a few
miscellaneous ones (Might need to do some customization to make them
work since there was literally 4 doors of on type that had no open doors
or anything).
- Portcullis and sliding doors will have their own files.