Container moves and changes, Planter deco's added
- Rescripted most basic utility containers, though the weights are set
to 255 on some more often used. This is due to world placement in static
development and from TileData settings. To override this, BaseWeight can
be used (In crafting creation) or a new Definition made. If this becomes
a huge inconvenience, I will change to crafted/standard weights instead
of TileData weights.
- Coded the planters from the 13th/14th anniversary gifts. They are just
decorative in nature.