Cooperative Collection, Alchemist, Tools
- Moved: Mortar and Pestle are now in a Tools.scp file in the Utility
folder.
- Changed: Updated some of the Axis Cats and fixed a redundancy/typo in
Alchemy items. Soon these will be moved to Decorations folder.
- Added the Cooperative Collection file. So far it has the Reading
Lantern. This comes in 5 hues which still need to be grabbed. When it
gets closer to time to work on the collection code, I will decide
whether to just do the color assignment via the Reward selection or to
just script up variants of the item assets in different hues.