Update 166: Barding Diff test, UOR Magic Item Generator, New Flags and more. Read notes.
- Added: UO:R/LBR Magic Item Generator. This file will test for karma on
creatures and generate the old magic items (Hence my removal of them).
It assigns a new prop, flags, and does checks to assign property
increases to the item itself.
- Added: Barding Difficulty test code: This is the true formula for
barding difficulty. It still needs testing on the math and decimal
trimming, but there should never be a preset barding difficulty tag on
anything. It is generated.
- Updated: Sound definitions got a minor update in naming conventions.
- Updated: HTMLDefinitions got a half way decent update with font color
selections. There is now a list containing all 140 CSS named web
compliant standard colors recognized by the W3 and browsers.
- Added: Recharge function for the item prop (Pending change in the new
experimental server)
- Added: Global behavior for the NPCs to generate Barding Difficulty
property.
- Added: tScroll now has its own CliLocs.
- Added: Pedestals art assets to decorations folder.
- Added: Coffee plant and code for the plants growth cycles and
harvesting.
- Added: Bar Sign to 17th anniversary items. Activated on DClick.
- Removed: Tools file
- Moved: Move tools to independent files for their skill categories.
- Added: Steam Powered Beverage maker and grinder. Grinder can be
DClicked on and targets Coffee Pods to make coffee grounds. At the
moment I am not using proper color or asset for grounds. It is a place
holder.
- Updated: Cleaned up the NPC files a bit more and removed code
generating 0 in properties. Sounds got added to quite a few as well.