Update 171: Rather large. Read notes.
- Updated: Barding Difficulty now is multiplied by 10 to generate a
decimal value, this tag will be changed to DIFFICULTY= as a prop in the
source to do more valid checks for skill gains as well on other skills.
- Updated: Attr - Added splintering weapon and paragon chest.
- Fixed: Typos in BuffClilocs from the old sphere scripts. They were
parsing wrong and using the same clilocs in two instances creating
issues.
- Fixed: Typo in BuffIcons from the original packs (They let a lot slip
^_^)
- Added: New CanI flags for Commodity, Deedable. Changed Powder to
fortify with a backwards compat until I switch the other ones over.
- Added: Def race  index. Useful for racial hatreds and warring.
- Added: Dye Flags - BoatDye to the dye indexes.
- Updated: Hues - Added another spelling constant for Light Grey.
- Added: Layers - Added LayerDeathPack, LayerFlagDisguise,
LayerspellAnimalForm.
- Added: Region Flags - Added NoLogin, NoMounts, NoPotions, NoPets,
NoSpeech, NoSpells, NoCombat, AntuMagicMark.
- Added: StatFlags - StatFDisguised, StatFDisarmed, StatFBleeding,
StatFMortalWound, StatFDismounted.
- Updated: TileFlags - Cleaned these up a bit using a more solid #
structure on the TFlagUnknowns.
- Updated: TypeDefinitions - Removed tArcheryWeapon and added a
backwards copat for tHide.
- Added: AIEvents - New AISavageWarrior event. Fixed the Daemon Bone
issue with the improper spacing in the <R<Val.Min>,<Val.Max>
- Updated: HumanEvents - Removed the unneeded Human prefixing as these
can be used on any Non-Hostile humanoid mobile.
- Updated: NPCEvents - Removed the Carnivore#'s and made one carnivore
events. It now processes logics for hunting and running, with checks for
food and other logics. This has a few other functions added atm I need
to move to other files. Terathan and ophidian now have checks to fight
each other as well. This should be moved to the AI files in the future.
Added instinct code to horse and ostards. Needs to be moved to AI
coding.
- Updated: AICombatFunctions - PetHardred is now IncreasePetDamage which
runs by a percentage. ATM using 100 in this will just do normal damage.
It can essentially reduce or increase damages. This might get a minor
rework to not need the 100 base. But for now it is fine.
- Added: CliLocs functions - Updated the files some with Extreme's
contributions and some minor cleanups.
- Updated: CoreFunctionsc - Cleaned up expansion check functions some.
Added SayL and MsgL. These are just faster ways to process CliLoc Says
and Messages by just using SayL <Color you want text to be>,<CliLoc val>
- Changed: cPropFunctions - Alignment is now FightMode since using
Alignment COULD cause issues with source if done as Alignments. So we
are trying to avoid any issues with a new naming structure.
- Updated: IsFunctions - Changed most of the QVal checks to EVal, QVal
could be a bit much for a simple check since it is more used for
complicated comparisons with different variant outcomes in one
structure. The EVal structure alone will still parse 1 if True and 0 if
False.
- Added: GlobalItemBehavior - Added in new checks for Tag.Override.Speed
to carry over on weapons. Honestly, I was too tired when I added it to
see if it carried over already, but it is a good safe guard atm. Also,
DClicking a weapon will not equip it, as sphere normally allows. In fact
this needs more updating for things that have internal targeting
functions.
- Added: GlobalNPCBehavior - Changed @BardingEffect to @BardEffect
(Could be confused for animal barding being added). Changed CanUse order
logics. Added in Tag.Override.Speed checks in @HitTry for mobiles.
Defaulting to 20 (2 seconds) if not declared for unarmed fighting. The
@Hit now checks for a server def for Pre-AoS Damage. This will lower
Dam.Lo by 1/2 and raised Dam.Hi by *2. like old UO:R. Added in
@ReceiveItem checks to stop wasting food units on pets. They will take
their maximum and refuse any more extras.
- Added: GlobalPlayerBehavior: As above for everything minus the Food.
Also added in Searing Weapon Code.
- Added: iPropfunctions - CarryDamTypes, CarryResTypes, CarryAbilities.
Needs a further check added, but these are now carrying over resistances
and such from original items on new creates. Need to add a check for
Magic and Artifact attr to not carry over on these in case they are hand
coded.
- Added: IsFunctions - Added IsSavage, IsDeedable, IsCommodity
- Fixed: TypeDefs- Fixed sphere from assigning a creature define in
More2 on leather and hides.
- Added: Just a list instead of writing each file: Dressforms, Knitting
tools, Stretched Hides, Ball of yarn, bolt of cloth (Several), Furs,
pelts,, cloth, cut cloth, cut leather, hides,  folded cloth, spool of
thread, Pile of wool, etc (Basically almost all of the tailoring items).
- Added all known bears, Added cu sidhe with proper code (Need to test
@Mount on them though, if it is not firing on them, we need to change
source to do so. Because @DClick shouldn't be used for this right now
for various reasons. Also, @Mount shouldn't be called from a player
since it would fire every on everything they technically try to DClick)
(EXTREME TEST THIS).
- Added: New myrmidex anims but we dont have the info on the new ones
for stats yet (Cough Nyx).
- Added: Rat, sewer rat, prison rat, giant rat, squirrel (sable and the
albino), ferrets, rabbit, jack rabbit, fire bunny, vorpal bunny.
- added: Original savage anims to base bodies for expansion into player
use :D and for further designs for the new tribes of Eodon.
- Added: Savage warriors, rider, and Shaman. Shaman is lacking the AI
coding atm since I need to test it more on my personal server to create
a more basic one for packs versus what we use on my server.
- Added: Light daemon, void daemon, and Djin (Nyx, I need to know what
this in CC, because the anim for EC was a bit unknown to me and didn't
match with anything on sites. I assumed the Chaos Daemon, but it looks
more like a ravager/moloch mix, but those didnt match up either from
reviewing assets).
- Added: ALL OF THE NON-BOSS DRAGONS (Thanks DarkFae for the
contribution)! Sorry for caps, but editing afterwards took a bit of
time, as I am teaching DarkFae the basics for NPC coding. :) It surely
was a lot of dragons though (Good job!)
- Updated: All Mobile files got some props moved to main blocks and had
their MaxFood moved to @Create.
- Updated: Savage Names - Added a few more to increase the #'s some.
Though, it seems both male and female names are the same on EA (Which is
odd, but whatev's).
- Updated: Lumberjacking code in Skills.scp - YOU SHALL NOT GET LOGS
WITH A SWORD (Cry about it, it is not proper, doesn't match EA and
defies all logics) Use an AXE (Like you are supposed to). Updated other
code in the skill too to make it work more properly.
- Updated: Spawns - Cleaned up some naming and added in the savage
spawns.
- Added: Test.scp for the new mobiles added that are uncategorized from
recent publish that were not myrmidex.
- Updated: Regions.scp - Fixed the Definition Naming for the new ores in
the files (Xun, Tag.Music was called in here oddly enough (I guess as an
override in the map files, but I don't even see why and might remove
it.)).
added new tools - Scissors, sewiing kit, dyes, dye vat, spinning wheels,
elven spinning wheels, looms, loom bench (Fixex the can't sit issue, but
I see WHY the did it (Notice some of the looms have it built into the
art disallowing the seating)).
Note: Source was updated for the elven spinning wheel to work more
accurately, since the old code would have broken its animation).
- Updated: Armor  - Again, moved resistances to main block. Added in
CliLoc naming.